move endpoint_slice functions to their own file
